requirements.txt
uvicorn
python-dotenv
httpx
fastapi[all]
openai
ngrok